refactor: linting and refactoring utils.py
